adding/updating tests

+++ packages/core/src/lib/assets/createFilesContext.ts
@@ -44,14 +44,10 @@ const generateSafeFilename = (
 export function createFilesContext(
   config: KeystoneConfig,
   s3Assets: () => Map<string, AssetsAPI>
-): FilesContext | undefined {
-  if (!config.storage) {
-    return;
-  }
-
+): FilesContext {
   const { storage } = config;
 
-  Object.entries(storage).forEach(([, val]) => {
+  Object.entries(storage || {}).forEach(([, val]) => {
     if (val.type === 'file' && val.kind === 'local') {
       fs.mkdirSync(val.storagePath || DEFAULT_FILES_STORAGE_PATH, { recursive: true });
     }
@@ -135,16 +131,24 @@ export function createFilesContext(
         case 's3': {
           const s3Instance = s3Assets().get(storageString);
 
+          if (!s3Instance) {
+            throw new Error(`Keystone has no connection to S3 storage location ${storageString}`);
+          }
+
           await s3Instance?.files.delete(filename);
           break;
         }
         case 'local': {
-          // TODO why is this not narrowing
           await fs.remove(
             path.join(storageConfig.storagePath || DEFAULT_FILES_STORAGE_PATH, filename)
           );
+          return;
         }
       }
+
+      throw new Error(
+        `attempted to get URL for storage ${storageString}, however could not find the config for it`
+      );
     },
   };
 }
